backend/requirements.txt
fastapi
uvicorn
moviepy
pytube
openai-whisper
numpy
Pillow
yt-dlp
webvtt-py